flags to byte complete and test

+++ src/main/java/kostiskag/unitynetwork/bluenode/functions/HashFunctions.java
@@ -70,7 +70,7 @@ public class HashFunctions {
     public static byte buildByteFromBits(String flags) {
     	byte b = (byte) 0x00;
     	if (flags.length() == 8) {
-    		b = Byte.parseByte(flags, 2);
+    		b = (byte) Integer.parseInt(flags, 2);
     	}
     	return b;
     }
